test: update tests for asserts of error message

+++ src/common/error/src/ext.rs
@@ -42,7 +42,7 @@ pub trait ErrorExt: StackError {
         match self.status_code() {
             StatusCode::Unknown | StatusCode::Unexpected | StatusCode::Internal => {
                 // masks internal error from end user
-                format!("Internal error: {}", self.status_code())
+                format!("Internal error: {}", self.status_code() as u32)
             }
             _ => {
                 let error = self.last();
